+++ packages/elements/src/color-picker/__test__/color-picker.test.js
@@ -3,6 +3,11 @@ import { fixture, expect, elementUpdated, keyboardEvent, oneEvent } from '@refin
 import '@refinitiv-ui/elements/color-picker';
 import '@refinitiv-ui/elemental-theme/light/ef-color-picker';
 
+/**
+ * Get private dialog element property
+ */
+export const getDialogEl = colorPicker => colorPicker.dialogRef.value;
+
 describe('color-picker/ColorPicker', () => {
 
   describe('DOM structure', () => {
@@ -25,6 +30,13 @@ describe('color-picker/ColorPicker', () => {
       const el = await fixture('<ef-color-picker value="#001EFF"></ef-color-picker>');
       expect(el.value).to.equal('#001EFF');
     });
+    it('Should update aria-label when change value color', async () => {
+      const el = await fixture('<ef-color-picker value="#0000FF"></ef-color-picker>');
+      expect(el.colorAriaLabel).to.equal('blue');
+      el.value = '#ff0000';
+      await elementUpdated(el);
+      expect(el.colorAriaLabel).to.equal('red');
+    });
     it('Should reset to default value when value is invalid', async () => {
       const el = await fixture('<ef-color-picker value="hello"></ef-color-picker>');
       expect(el.value).to.equal('');
@@ -43,7 +55,7 @@ describe('color-picker/ColorPicker', () => {
     });
     it('Should fires value-changed event when value change by user interactions', async () => {
       const el = await fixture('<ef-color-picker value="#001EFF" opened></ef-color-picker>');
-      const dialogEl = el.dialogEl;
+      const dialogEl = getDialogEl(el);
       const redInput =  dialogEl.shadowRoot.getElementById('redInput');
       const confirmBtn = dialogEl.shadowRoot.getElementById('confirmButton');
       redInput.value = 200;
@@ -61,13 +73,13 @@ describe('color-picker/ColorPicker', () => {
       const el = await fixture('<ef-color-picker></ef-color-picker>');
       el.opened = true;
       await elementUpdated(el);
-      expect(el.dialogEl.hasAttribute('allow-nocolor')).to.be.equal(false);
+      expect(getDialogEl(el).hasAttribute('allow-nocolor')).to.be.equal(false);
     });
     it('Should pass allow-nocolor property to color dialog', async () => {
       const el = await fixture('<ef-color-picker allow-nocolor></ef-color-picker>');
       el.opened = true;
       await elementUpdated(el);
-      expect(el.dialogEl.hasAttribute('allow-nocolor')).to.be.equal(true);
+      expect(getDialogEl(el).hasAttribute('allow-nocolor')).to.be.equal(true);
     });
   });
 
@@ -76,13 +88,13 @@ describe('color-picker/ColorPicker', () => {
       const el = await fixture('<ef-color-picker></ef-color-picker>');
       el.click();
       await elementUpdated(el);
-      expect(el.dialogEl.opened).to.be.equal(true, 'clicking on color picker should open color dialog');
+      expect(getDialogEl(el).opened).to.be.equal(true, 'clicking on color picker should open color dialog');
     });
     it('Should open dialog when opened programmatically', async () => {
       const el = await fixture('<ef-color-picker></ef-color-picker>');
       el.opened = true;
       await elementUpdated(el);
-      expect(el.dialogEl.hasAttribute('opened')).to.be.equal(true);
+      expect(getDialogEl(el).hasAttribute('opened')).to.be.equal(true);
     });
     it('Should not open color dialog when disabled', async () => {
       const el = await fixture('<ef-color-picker disabled></ef-color-picker>');
@@ -101,13 +113,13 @@ describe('color-picker/ColorPicker', () => {
       const el = await fixture('<ef-color-picker></ef-color-picker>');
       el.dispatchEvent(keyboardEvent('keydown', { key: 'Enter' }));
       await elementUpdated(el);
-      expect(el.dialogEl.opened).to.be.equal(true, 'Enter should open dialog');
+      expect(getDialogEl(el).opened).to.be.equal(true, 'Enter should open dialog');
     });
     it('Should open dialog when press spacebar key', async () => {
       const el = await fixture('<ef-color-picker></ef-color-picker>');
       el.dispatchEvent(keyboardEvent('keydown', { key: 'Spacebar' }));
       await elementUpdated(el);
-      expect(el.dialogEl.opened).to.be.equal(true, 'Spacebar should open dialog');
+      expect(getDialogEl(el).opened).to.be.equal(true, 'Spacebar should open dialog');
     });
   });
 });