+++ packages/phrasebook/src/locale/de/color-picker.ts
@@ -1,9 +1,22 @@
-// Component docs https://elf.int.refinitiv.com/elements/color-picker.html
+// Component docs https://ui.refinitiv.com/elements/color-picker
 import { Phrasebook } from '../../translation.js';
 import './shared.js';
 import './color-dialog.js';
 
 const translations = {
+  VERY_LIGHT: 'Sehr hell',
+  LIGHT: 'Hell',
+  VERY_DARK: 'Sehr dunkel',
+  DARK: 'Dunkel',
+  BLACK: 'Schwarz',
+  WHITE: 'Weiß',
+  WITH: 'Mit {number}%',
+  YELLOW: 'Gelb',
+  GREEN: 'Grün',
+  CYAN: 'Cyan',
+  BLUE: 'Blau',
+  MAGENTA: 'Magenta',
+  RED: 'Rot'
 };
 
 Phrasebook.define('de', 'ef-color-picker', translations);