+++ packages/theme-compiler/src/helpers.js
@@ -7,8 +7,20 @@ const autoprefixer = require('postcss')().use(require('autoprefixer'));
 const clean = new (require('clean-css'))({ returnPromise: true, level: '2' });
 const path = require('path');
 
-const wrap = (name, style) => `elf.${name.indexOf('-') > 0 ? 'custom'
-: 'native'}Styles.define('${name}', '${style.replace(/'/g, '\\\'')}');\n`;
+/**
+ * Return injector code in form of string
+ * @param {string} name element's path
+ * @param {string} style element's path
+ * @param {string} isEvent condition if need to be event method
+ * @returns {string} injector code
+ */
+const wrap = (name, style, isEvent) => {
+  const eventName = name.indexOf('-') > 0 ? 'custom' : 'native';
+  if(isEvent) {
+    return `dispatchEvent(new CustomEvent('ef.${eventName}Styles.define', { detail: { name: '${name}', styles: '${style.replace(/'/g, '\\\'')}' }}));\n`;
+  }
+  return `elf.${eventName}Styles.define('${name}', '${style.replace(/'/g, '\\\'')}');\n`;
+}
 
 const cleanCSS = css => autoprefixer.process(css, { from: false })
 .then(o => clean.minify(o.css)).then(o => o.styles);
@@ -16,6 +28,13 @@ const cleanCSS = css => autoprefixer.process(css, { from: false })
 const wrapHostSelectors = css => Promise
 .resolve(css.replace(/(:host)([.:[#][^\s,{]+)/g, '$1($2)'));
 
+/**
+ * Return less option template
+ * @param {string} entrypoint source map output
+ * @param {string} filename source of the element styles
+ * @param {string} variables variables that override the less file
+ * @returns {string} injector code
+ */
 const generateLessOptions = (entrypoint, filename, variables) => ({
   filename: entrypoint,
   math: 2,
@@ -41,13 +60,21 @@ const generateLessOptions = (entrypoint, filename, variables) => ({
   ]
 });
 
-const generateJsInfo = (name, css, dependencies) => {
+/**
+ * Return generated info for injector string
+ * @param {string} name element name
+ * @param {string} css element style
+ * @param {string} dependencies list of elements
+ * @param {string} variables option variables that include using event condition
+ * @returns {object} injector code
+ */
+const generateJsInfo = (name, css, dependencies, variables) => {
   let importString = 'import \'./imports/native-elements.js\';\n';
   importString += dependencies.filter(name => name.indexOf('-') !== -1)
   .map(dep => `import './${dep}.js';`).join('\n') + '\n';
   return {
     importString,
-    injectorString: wrap(name, css.replace(/([^\\])\\([^\\])/g, '$1\\\\$2'))
+    injectorString: wrap(name, css.replace(/([^\\])\\([^\\])/g, '$1\\\\$2'), variables.registration === 'event')
   };
 };
 
@@ -59,13 +86,20 @@ const getElementNameFromLess = (filename) => {
   return path.basename(filename).replace(/\.less$/, '');
 };
 
-const generateOutput = (filename, output) => {
+/**
+ * Return object that use for parser
+ * @param {string} filename element source
+ * @param {string} output less file source 
+ * @param {string} variables option variables that include using event condition
+ * @returns {object}
+ */
+const generateOutput = (filename, output, variables) => {
   return cleanCSS(output.css).then(wrapHostSelectors).then(css => {
     let name = path.basename(filename).replace(/\.less$/, '');
     let dependencies = output.imports
     .filter(filename => prefix.test(filename))
     .map(filename => filename.replace(dependencyPattern, ''));
-    let styleInfo = generateJsInfo(name, css, dependencies);
+    let styleInfo = generateJsInfo(name, css, dependencies, variables);
     return {
       name,
       dependencies,