+++ packages/i18n/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.9](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/i18n@6.0.7...@refinitiv-ui/i18n@6.0.9) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/i18n
+
+
+
+
+
+## [6.0.8](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/i18n@6.0.7...@refinitiv-ui/i18n@6.0.8) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/i18n
+
+
+
+
+
 ## [6.0.7](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/i18n@6.0.6...@refinitiv-ui/i18n@6.0.7) (2022-10-26)
 
 **Note:** Version bump only for package @refinitiv-ui/i18n