+++ packages/test-helpers/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.6](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/test-helpers@6.0.5...@refinitiv-ui/test-helpers@6.0.6) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/test-helpers
+
+
+
+
+
+## [6.0.5](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/test-helpers@6.0.4...@refinitiv-ui/test-helpers@6.0.5) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/test-helpers
+
+
+
+
+
 ## [6.0.4](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/test-helpers@6.0.3...@refinitiv-ui/test-helpers@6.0.4) (2022-08-29)
 
 **Note:** Version bump only for package @refinitiv-ui/test-helpers