+++ packages/translate/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.11](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/translate@6.0.9...@refinitiv-ui/translate@6.0.11) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/translate
+
+
+
+
+
+## [6.0.10](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/translate@6.0.9...@refinitiv-ui/translate@6.0.10) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/translate
+
+
+
+
+
 ## [6.0.9](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/translate@6.0.8...@refinitiv-ui/translate@6.0.9) (2022-10-26)
 
 **Note:** Version bump only for package @refinitiv-ui/translate