+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.2.2](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/utils@6.2.1...@refinitiv-ui/utils@6.2.2) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/utils
+
+
+
+
+
+## [6.2.1](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/utils@6.2.0...@refinitiv-ui/utils@6.2.1) (2022-11-07)
+
+**Note:** Version bump only for package @refinitiv-ui/utils
+
+
+
+
+
 # [6.2.0](https://github.com/Refinitiv/refinitiv-ui/compare/@refinitiv-ui/utils@6.1.1...@refinitiv-ui/utils@6.2.0) (2022-10-26)